Rename slack-moderate-words to slack-moderator-words
