--- slack-moderate-words/Dockerfile
@@ -1,3 +1,0 @@
-FROM gcr.io/distroless/base
-ADD slack-moderate-words /slack-moderate-words
-ENTRYPOINT ["/slack-moderate-words"]
