+++ slack-moderator-words/.gitignore
@@ -1,4 +1,4 @@
 config.json
 
 # go binary
-slack-moderate-words
+slack-moderator-words
