+++ slack-moderator-words/Dockerfile
@@ -1,0 +1,3 @@
+FROM gcr.io/distroless/base
+ADD slack-moderator-words /slack-moderator-words
+ENTRYPOINT ["/slack-moderator-words"]
