+++ slack-moderator-words/filters.yaml
