+++ slack-moderator-words/main.go
@@ -27,7 +27,7 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"sigs.k8s.io/slack-infra/slack"
-	"sigs.k8s.io/slack-infra/slack-moderate-words/model"
+	"sigs.k8s.io/slack-infra/slack-moderator-words/model"
 )
 
 type options struct {
