+++ slack-moderator-words/model/types.go
